requirements.txt
gmusicapi
pyyaml